backend/requirements.txt
fastapi
uvicorn[standard]
rio-tiler
rasterio
Pillow
python-dotenv
